feat: Add dashboard route to public routes
The middleware.ts file was modified to include the '/dashboard(.*)' route in the list of public routes. This change allows unauthenticated users to access the dashboard page.

.env.local file was modified to add force redirect URLs for sign-in and sign-up. The URLs were set to '/' to redirect users to the home page after signing in or signing up.

A new page component, update-info/page.tsx, was added under the src/app/(booking) directory. This page component renders the UpdateStaffInfo component.

The db.ts file was modified to export a singleton instance of the PrismaClient. This ensures that only one instance of the PrismaClient is created and used throughout the application.

The page component update-info/page.tsx under the src/app/(auth)/(routes) directory was deleted. This page component is no longer needed and was removed from the codebase.

The compose.yaml file was modified to remove the app service configuration. The app service was no longer needed and was removed from the compose file.

The package.json file was modified to update the version of the @prisma/client dependency to 5.21.1.

The HomeCleaning component in src/components/step-1/HomeCleaning.tsx was modified to add the addBookingDetail and addServiceCategory actions. These actions are called when the "Next" button is clicked.

The actions.ts file was modified to add the addBookingDetail and addServiceCategory functions. These functions are used to add booking details and service categories to the database.

The UpdateStaffInfo component in src/components/updateinfo/staff/UpdateStaffInfo.tsx was modified to include form data handling and submission. The form data is stored in the component's state and the addHelper action is called when the "Verify" button is clicked.